Move computation functions into respective modules.
